requirements_old.txt
absl-py==2.1.0
anyio==4.2.0
apptools==5.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==4.1.0
Brotli==1.0.9
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorcet==3.1.0
comm==0.2.1
connected-components-3d==3.18.0
contourpy==1.1.1
cycler==0.12.1
cycpd==0.28
Cython==3.0.11
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dosma==0.1.2
envisage==7.0.3
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==0.9.1
fastjsonschema==2.16.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.0
huggingface-hub
h11==0.14.0
h5py==3.11.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
imageio==2.35.1
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipydatawidgets==4.3.2
ipykernel==6.15.1
ipympl==0.9.3
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.8.3
isodate==0.6.1
itk-core==5.4.0
itk-filtering==5.4.0
itk-meshtopolydata==0.11.0
itk-numerics==5.4.0
itkwidgets==0.32.6
jedi==0.18.1
Jinja2==3.1.4
json5==0.9.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
jupyterlab_widgets==1.1.9
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.41.1
looseversion==1.3.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.3
mayavi==4.8.2
mistune==2.0.4
mskt==0.0.10
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.5
nested-lookup==0.2.25
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
notebook==6.4.13
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.4.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.0
param==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
ply==3.11
Pmw==2.1.1
point-cloud-utils==0.30.4
pooch==1.8.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==4.25.4
prov==2.0.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyacvd==0.2.11
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydicom==2.3.0
pydot==3.0.1
pyface==8.0.0
pyfocusr==0.0.15
Pygments==2.15.1
pymeshfix==0.16.3
pyparsing==3.1.4
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyvista==0.44.1
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==25.1.2
qtconsole==5.6.0
QtPy==2.4.1
rdflib==6.3.2
referencing
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
scooby==0.10.0
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==72.1.0
sigpy==0.1.26
SimpleITK==2.4.0
simplejson==3.19.3
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.3.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.17.1
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.66.5
traitlets
traits==6.3.2
traitsui==8.0.0
traittypes==0.2.1
typing_extensions
tzdata==2024.1
urllib3==2.2.2
vtk==9.3.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.43.0
widgetsnbextension==3.6.8
wrapt==1.16.0
zipp==3.20.1
zstandard==0.23.0
